Converting custom expectations in it own type
